refactor(backend/sdoc_source_code): de-duplicate line_marker_processor
This change makes both language-aware and unaware parsers to use the same line marker validation function.

This is a follow-up to: #2458
