+++ strictdoc/backend/sdoc_source_code/reader.py
@@ -3,7 +3,7 @@
 """
 
 from functools import partial
-from typing import Any, Dict, List, Optional, TypedDict, Union
+from typing import Any, Dict, List, Optional, TypedDict
 
 from textx import get_location, metamodel_from_str
 
@@ -21,8 +21,11 @@ from strictdoc.backend.sdoc_source_code.models.source_file_info import (
     SourceFileTraceabilityInfo,
 )
 from strictdoc.backend.sdoc_source_code.parse_context import ParseContext
+from strictdoc.backend.sdoc_source_code.processors.general_language_marker_processors import (
+    line_marker_processor,
+    validate_marker_uids,
+)
 from strictdoc.helpers.file_stats import SourceFileStats
-from strictdoc.helpers.list import find_duplicates
 from strictdoc.helpers.textx import drop_textx_meta
 
 
@@ -137,20 +140,6 @@ def create_unmatch_range_error(
     )
 
 
-def validate_marker_uids(
-    marker: Union[FunctionRangeMarker, LineMarker, RangeMarker],
-    parse_context: ParseContext,
-) -> None:
-    possible_duplicates = find_duplicates(marker.reqs)
-    if len(possible_duplicates) > 0:
-        location = get_location(marker)
-
-        raise ValueError(
-            "@relation marker contains duplicate node UIDs: "
-            f"{possible_duplicates}. Location: {parse_context.filename}:{location['line']}."
-        )
-
-
 def range_marker_processor(
     marker: RangeMarker, parse_context: ParseContext
 ) -> None:
@@ -210,57 +199,6 @@ def range_marker_processor(
             raise create_end_without_begin_error(location) from None
     else:
         raise NotImplementedError
-
-
-def line_marker_processor(
-    line_marker: LineMarker, parse_context: ParseContext
-) -> None:
-    validate_marker_uids(line_marker, parse_context)
-
-    location = get_location(line_marker)
-    line = location["line"]
-
-    if (
-        len(parse_context.marker_stack) > 0
-        and parse_context.marker_stack[-1].ng_is_nodoc
-    ):
-        # This marker is within a "nosdoc" block, so we ignore it.
-        return
-
-    # Semantic constraint validation
-    has_previous_markers = len(parse_context.markers) > 0
-    is_consecutive = (
-        has_previous_markers
-        and parse_context.markers[-1].ng_range_line_end == line
-    )
-    is_at_eof = line == parse_context.file_stats.lines_total
-
-    if is_consecutive:
-        raise StrictDocSemanticError(
-            "Consecutive LineMarkers are not allowed",
-            hint=None,
-            example=None,
-            line=line,
-            filename=location["filename"],
-        )
-
-    if is_at_eof:
-        raise StrictDocSemanticError(
-            "LineMarker cannot be followed by EOF",
-            hint=None,
-            example=None,
-            line=line,
-            filename=location["filename"],
-        )
-
-    parse_context.markers.append(line_marker)
-    line_marker.ng_source_line_begin = line
-    line_marker.ng_range_line_begin = line
-    line_marker.ng_range_line_end = line + 1
-
-    for req in line_marker.reqs:
-        markers = parse_context.map_reqs_to_markers.setdefault(req, [])
-        markers.append(line_marker)
 
 
 def function_range_marker_processor(
