+++ tests/integration/features/source_code_traceability/_general_parser_backward/50_line_marker/file.py
@@ -1,7 +1,7 @@
 def hello_world():
-    """
-    @sdoc(REQ-001)
-    """
+    # @relation(REQ-001, scope=line)
+    print("Line marker")  # noqa: T201
+
     # @sdoc[REQ-001]
     print("ignored hello world")  # noqa: T201
     print("ignored hello world")  # noqa: T201
