+++ tests/integration/features/source_code_traceability/_general_parser_backward/50_line_marker/test.itest
@@ -6,5 +6,5 @@ CHECK: Published: Hello world doc
 RUN: %check_exists --file "%T/html/_source_files/file.py.html"
 
 RUN: %cat %T/html/_source_files/file.py.html | filecheck %s --dump-input=fail --check-prefix CHECK-SOURCE-FILE
-CHECK-SOURCE-FILE: [ 3 ]
+CHECK-SOURCE-FILE: [ 2-3 ]
 CHECK-SOURCE-FILE: [ 5-8 ]
