+++ tests/integration/features/source_code_traceability/_language_parsers/python/03_python_forward_function/test.itest
@@ -11,12 +11,12 @@ RUN: %cat %T/html/%THIS_TEST_FOLDER/input.html | filecheck %s --dump-input=fail 
 CHECK-HTML: <a{{.*}}href="../_source_files/file.py.html#REQ-1#1#10">
 CHECK-HTML: <a{{.*}}href="../_source_files/file.py.html#REQ-1#1#10">
 CHECK-HTML: <a{{.*}}href="../_source_files/file.py.html#REQ-1#5#10">
-CHECK-HTML: <a{{.*}}href="../_source_files/file.py.html#REQ-1#7#7">
+CHECK-HTML: <a{{.*}}href="../_source_files/file.py.html#REQ-1#7#8">
 
 RUN: %cat %T/html/_source_files/file.py.html | filecheck %s --dump-input=fail --check-prefix CHECK-SOURCE-FILE
 CHECK-SOURCE-FILE: 1 - 10 | function hello_world()
 CHECK-SOURCE-FILE: 5 - 10 | range
-CHECK-SOURCE-FILE: 7 - 7 | line
+CHECK-SOURCE-FILE: 7 - 8 | line
 
 RUN: %cat %T/html/source_coverage.html | filecheck %s --dump-input=fail --check-prefix CHECK-SOURCE-COVERAGE
 CHECK-SOURCE-COVERAGE: 83.3
